Complete final code standard check.
