+++ src/main/java/gui/Duke.java
@@ -3,16 +3,14 @@ package gui;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
-import javafx.scene.layout.Region;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-
 import windycall.WindyCall;
 
 
@@ -29,12 +27,12 @@ public class Duke extends Application {
 
     private WindyCall windycall;
 
-    public static void main(String[] args) {
-        // ...
-    }
-
     public Duke() {
         this.windycall = new WindyCall();
+    }
+
+    public static void main(String[] args) {
+        // ...
     }
 
     @Override
@@ -144,8 +142,5 @@ public class Duke extends Application {
     public String getResponse(String input) {
         return windycall.run(input);
     }
-
-
-
 
 }
