+++ src/main/java/windycall/handler/AddDeadlineHandler.java
@@ -1,11 +1,11 @@
-package windycall.Handler;
+package windycall.handler;
 
+import java.time.LocalDate;
+
+import windycall.exception.WindyCallException;
 import windycall.parser.Parser;
 import windycall.task.Deadline;
 import windycall.task.Task;
-import windycall.exception.WindyCallException;
-
-import java.time.LocalDate;
 
 public class AddDeadlineHandler extends AddTaskHandler {
 
@@ -27,7 +27,6 @@ public class AddDeadlineHandler extends AddTaskHandler {
             throw new WindyCallException("OOPS!!! You haven't specify the deadline for the Task");
         }
         String description = message.substring(9, idx - 1);
-//        System.out.println("Got it. I've added this Deadline task:");
         String deadlineStr = message.substring(idx + 4);
         LocalDate deadline = Parser.processDate(deadlineStr);
         return new Deadline(description, false, deadline);
