+++ src/main/java/windycall/handler/DeleteOperationHandler.java
@@ -1,13 +1,13 @@
-package windycall.Handler;
+package windycall.handler;
+
+import java.util.List;
 
 import javafx.util.Pair;
 import windycall.parser.Parser;
 import windycall.storage.Storage;
 import windycall.task.Task;
 
-import java.util.List;
-
-public class AddTagHandler extends OperationHandler {
+public class DeleteOperationHandler extends OperationHandler {
 
     @Override
     public String handle(String userCommand) {
@@ -15,21 +15,21 @@ public class AddTagHandler extends OperationHandler {
     }
 
     public String handle(Parser parser, List<Task> tasks, String[] parts, Storage storage) {
-        // valid format: tag #index #tag
-        Pair<Integer, String> info = parser.getTagIndex(parts);
+        Pair<Integer, String> info = parser.getDeleteIndex(parts);
         int num = info.getKey();
-        String tag = info.getValue();
-//        assert num >= 1 && num <= tasks.size();
+        String message = info.getValue();
+        assert num >= 1 && num <= tasks.size();
         if (num >= 1 && num <= tasks.size()) {
-            tasks.get(num - 1).changeTag(tag);
-            String returnedMessage = "I've successfully added tag #" + tag + " to this task:\n";
-            returnedMessage += tasks.get(num - 1);
+            String returnedMessage = "Noted. I've removed this task:\n";
+            returnedMessage += tasks.get(num - 1) + "\n";
+            tasks.remove(num - 1);
+            returnedMessage += "Now you have " + tasks.size() + " tasks in the list.";
             storage.handleTaskChange(tasks);
             return returnedMessage;
         } else if (num > tasks.size() || (num < 1 && num != -1)) {
             return "Sorry, your index is out of range";
         } else {
-            return tag;
+            return message;
         }
     }
 }
