+++ src/main/java/windycall/handler/MarkOperationHandler.java
@@ -1,13 +1,13 @@
-package windycall.Handler;
+package windycall.handler;
+
+import java.util.List;
 
 import javafx.util.Pair;
 import windycall.parser.Parser;
 import windycall.storage.Storage;
 import windycall.task.Task;
 
-import java.util.List;
-
-public class DeleteOperationHandler extends OperationHandler {
+public class MarkOperationHandler extends OperationHandler {
 
     @Override
     public String handle(String userCommand) {
@@ -15,15 +15,14 @@ public class DeleteOperationHandler extends OperationHandler {
     }
 
     public String handle(Parser parser, List<Task> tasks, String[] parts, Storage storage) {
-        Pair<Integer, String> info = parser.getDeleteIndex(parts);
+        Pair<Integer, String> info = parser.getMarkIndex(parts);
         int num = info.getKey();
         String message = info.getValue();
         assert num >= 1 && num <= tasks.size();
         if (num >= 1 && num <= tasks.size()) {
-            String returnedMessage = "Noted. I've removed this task:\n";
-            returnedMessage += tasks.get(num - 1) + "\n";
-            tasks.remove(num - 1);
-            returnedMessage += "Now you have " + tasks.size() + " tasks in the list.";
+            tasks.get(num - 1).markAsDone();
+            String returnedMessage = "Good job! I've marked this task as done:\n";
+            returnedMessage += tasks.get(num - 1);
             storage.handleTaskChange(tasks);
             return returnedMessage;
         } else if (num > tasks.size() || (num < 1 && num != -1)) {
@@ -32,4 +31,6 @@ public class DeleteOperationHandler extends OperationHandler {
             return message;
         }
     }
+
+
 }
