+++ src/main/java/windycall/handler/OperationHandler.java
@@ -1,4 +1,4 @@
-package windycall.Handler;
+package windycall.handler;
 
 public abstract class OperationHandler {
