+++ src/main/java/windycall/handler/UnmarkOperationHandler.java
@@ -1,27 +1,26 @@
-package windycall.Handler;
+package windycall.handler;
 
 import java.util.List;
+
 import javafx.util.Pair;
 import windycall.parser.Parser;
 import windycall.storage.Storage;
 import windycall.task.Task;
 
-public class MarkOperationHandler extends OperationHandler {
-
+public class UnmarkOperationHandler extends OperationHandler {
     @Override
     public String handle(String userCommand) {
         return "";
     }
 
     public String handle(Parser parser, List<Task> tasks, String[] parts, Storage storage) {
-//        int num = parser.getMarkIndex(parts);
-        Pair<Integer, String> info = parser.getMarkIndex(parts);
+        Pair<Integer, String> info = parser.getUnmarkIndex(parts);
         int num = info.getKey();
         String message = info.getValue();
         assert num >= 1 && num <= tasks.size();
         if (num >= 1 && num <= tasks.size()) {
-            tasks.get(num - 1).markAsDone();
-            String returnedMessage = "Good job! I've marked this task as done:\n";
+            tasks.get(num - 1).unmark();
+            String returnedMessage = "Good job! I've unmarked this task as not done yet:\n";
             returnedMessage += tasks.get(num - 1);
             storage.handleTaskChange(tasks);
             return returnedMessage;
@@ -31,6 +30,4 @@ public class MarkOperationHandler extends OperationHandler {
             return message;
         }
     }
-
-
 }
