+++ src/main/java/windycall/task/Event.java
@@ -1,7 +1,5 @@
 package windycall.task;
 
-import windycall.parser.Parser;
-
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
